requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
flask>=2.0.0
# 长连接方式（bot_server_ws.py）需要，若只用 Webhook 可注释掉
lark-oapi>=1.1.0
# xlsx/xlsm 单元格级变更对比（可选，不装则 xlsx 仅显示“文件变更”）
openpyxl>=3.1.0
